+++ tests/test_polygonrnn_model.py
@@ -97,7 +97,7 @@ class Test_PolygonRNNModel(BasicTestCase):
                     f"train_dataset.root_dir={polygon_rnn_root_dir}",
                     f"val_dataset.input_csv_path={csv_path}",
                     f"val_dataset.root_dir={polygon_rnn_root_dir}",
-                    "+pl_trainer.fast_dev_run=true",
+                    "++pl_trainer.fast_dev_run=true",
                 ],
             )
             trainer = train(cfg)
@@ -112,6 +112,7 @@ class Test_PolygonRNNModel(BasicTestCase):
                     f"train_dataset.root_dir={polygon_rnn_root_dir}",
                     f"val_dataset.input_csv_path={csv_path}",
                     f"val_dataset.root_dir={polygon_rnn_root_dir}",
+                    "++pl_trainer.fast_dev_run=true",
                 ],
             )
             trainer = train(cfg)