+++ tests/test_predict.py
@@ -50,7 +50,7 @@ from pytorch_segmentation_models_trainer.utils.os_utils import (
     create_folder,
     remove_folder,
 )
-from torchvision.datasets.utils import download_url
+from tests.mock_utils import create_dummy_checkpoint
 
 
 config_name_list = ["predict.yaml"]
@@ -103,11 +103,7 @@ class Test_Predict(unittest.TestCase):
         checkpoint_folder = create_folder(os.path.join(root_dir, "data", "checkpoints"))
         ckeckpoint_file_path = os.path.join(checkpoint_folder, file_name)
         if not os.path.isfile(ckeckpoint_file_path):
-            download_url(
-                url=pretrained_checkpoints_download_links[file_name.split(".")[0]],
-                root=checkpoint_folder,
-                filename=file_name,
-            )
+            create_dummy_checkpoint(ckeckpoint_file_path)
         return ckeckpoint_file_path
 
     def make_inference(self, sample, frame_field_model):